added Validation.scala.
